Refactor SOAPRequest to use new API::Reponse#xml method
